WIP: authorization service interface now independent of config files.
The code is still broken. I still need an interface for computer.
