+++ src/Lithnet.Laps.Web/Lithnet.Laps.Web/Authorization/IAuthorizationService.cs
@@ -11,11 +11,7 @@ namespace Lithnet.Laps.Web.Authorization
         /// </summary>
         /// <param name="currentUser">a user. FIXME: this should just be the user name.</param>
         /// <param name="computerName">name of the computer</param>
-        /// <param name="target">Target section in the web.config-file.
-        /// FIXME: This shouldn't be in this interface. But I can't leave it out
-        /// yet, because the code figuring out the target has way too many dependencies.
-        /// </param>
         /// <returns>An <see cref="AuthorizationResponse"/> object.</returns>
-        AuthorizationResponse CanAccessPassword(UserPrincipal currentUser, string computerName, TargetElement target = null);
+        AuthorizationResponse CanAccessPassword(UserPrincipal currentUser, string computerName);
     }
 }
