+++ src/Lithnet.Laps.Web/Lithnet.Laps.Web/Internal/EventIDs.cs
@@ -5,11 +5,18 @@ using System.Web;
 
 namespace Lithnet.Laps.Web
 {
+    /// <summary>
+    /// Some kind of return code.
+    ///
+    /// FIXME: I think EventIDs is a bad name for this class.
+    /// Can't we replace this by an enum?
+    /// </summary>
     internal static class EventIDs
     {
         public const int PasswordAccessed = 200;
         public const int UserAuthenticated = 201;
         public const int UserRequestedPassword = 202;
+        public const int UserAuthorizedForComputer = 203;
 
         public const int SsoIdentityNotFound = 400;
         public const int ComputerNotFound = 401;
